Import of RCAPUTO/POE-0.16 from CPAN.
gitpan-cpan-distribution: POE
gitpan-cpan-version:      0.16
gitpan-cpan-path:         RCAPUTO/POE-0.16.tar.gz
gitpan-cpan-author:       RCAPUTO
gitpan-cpan-maturity:     released

+++ t/01_sessions.t
@@ -1,5 +1,5 @@
 #!/usr/bin/perl -w
-# $Id: 01_sessions.t,v 1.13 2001/06/07 14:29:46 rcaputo Exp $
+# $Id: 01_sessions.t,v 1.15 2001/07/25 15:50:27 rcaputo Exp $
 
 # Tests basic compilation and events.
 
@@ -11,6 +11,7 @@ use TestSetup;
 # Turn on all asserts.
 #sub POE::Kernel::TRACE_DEFAULT  () { 1 }
 sub POE::Kernel::ASSERT_DEFAULT () { 1 }
+sub POE::Session::ASSERT_STATES () { 0 }
 use POE;
 
 ### Test parameters and results.
@@ -106,8 +107,14 @@ POE::Session->create
       sub {
         if ($_[HEAP]->{kills_to_go}--) {
           $_[KERNEL]->delay( fire_signals => 0.5 );
-          kill ALRM => $$;
-          kill PIPE => $$;
+          if ($^O eq 'MSWin32') {
+            $_[KERNEL]->signal( $_[KERNEL], 'ALRM' );
+            $_[KERNEL]->signal( $_[KERNEL], 'PIPE' );
+          }
+          else {
+            kill ALRM => $$;
+            kill PIPE => $$;
+          }
         }
         # One last timer so the session lingers long enough to catch
         # the final signal.
@@ -392,11 +399,17 @@ for (my $i=0; $i<$machine_count; $i++) {
 }
 
 # Were all the signals caught?
-print 'not ' unless $sigalrm_caught == $event_count;
-print "ok 11\n";
+if ($^O eq 'MSWin32') {
+  print "ok 11 # skipped: Windows doesn't seem to do signals\n";
+  print "ok 12 # skipped: Windows doesn't seem to do signals\n";
+}
+else {
+  print 'not ' unless $sigalrm_caught == $event_count;
+  print "ok 11\n";
 
-print 'not ' unless $sigpipe_caught == $event_count;
-print "ok 12\n";
+  print 'not ' unless $sigpipe_caught == $event_count;
+  print "ok 12\n";
+}
 
 # Did the postbacks work?
 print 'not ' unless $postback_test;
